Uprava menu a aplikacie

+++ Volby_Semest/Vysledok.cpp
@@ -10,7 +10,7 @@ Vysledok::Vysledok() :
 void Vysledok::vycisliRelativnyVysledok(int hlasyVsetkyStrany)
 {
 	(hlasyVsetkyStrany == 0) ? relativnyVysledok_ = 0.0 :
-		relativnyVysledok_ = (static_cast<double>(absolutnyVysledok_) / static_cast<double>(hlasyVsetkyStrany)) * 100.0;
+		relativnyVysledok_ = (static_cast<double>(absolutnyVysledok_) / hlasyVsetkyStrany) * 100.0;
 }
 
 void Vysledok::pridajPrefHlasy(int pocet)
